+++ example/android/app/src/main/kotlin/com/example/flutter_local_network_ios_example/MainActivity.kt
@@ -2,5 +2,4 @@ package com.example.flutter_local_network_ios_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()